cache improvements
added the ability to clear the entire cache and the recurring interval. cache methods now return the cache instead of voiding. pending updates can be manually pushed to the database using a new method, `writeAllPendingUpdates()`. setting a value now pushes it to the pending updates cache IF the `usePendingUpdates` param is set to `true`.